fix nightmare
