+++ src/lib/minions/functions/index.ts
@@ -8,7 +8,6 @@ import killableMonsters from '../data/killableMonsters';
 import { KillableMonster } from '../types';
 
 export { default as reducedTimeForGroup } from './reducedTimeForGroup';
-export { default as findMonster } from './findMonster';
 export { default as calculateMonsterFood } from './calculateMonsterFood';
 
 export type AttackStyles =
@@ -21,21 +20,15 @@ export type AttackStyles =
 export function resolveAttackStyles(
 	user: KlasaUser,
 	monsterID: number
-): [KillableMonster, Monster, AttackStyles[]] {
+): [KillableMonster | undefined, Monster | undefined, AttackStyles[]] {
 	const killableMon = killableMonsters.find(m => m.id === monsterID);
 	const osjsMon = Monsters.get(monsterID);
-	if (!killableMon) {
-		throw new Error(`Missing killableMonster for ${monsterID}::${typeof monsterID}`);
-	}
-	if (!osjsMon) {
-		throw new Error(`Missing osjsMon for ${monsterID}::${typeof monsterID}`);
-	}
 
 	// The styles chosen by this user to use.
 	let attackStyles = user.getAttackStyles();
 
 	// The default attack styles to use for this monster, defaults to shared (melee)
-	const monsterStyles = killableMon.defaultAttackStyles ?? [
+	const monsterStyles = killableMon?.defaultAttackStyles ?? [
 		SkillsEnum.Attack,
 		SkillsEnum.Strength,
 		SkillsEnum.Defence
@@ -45,7 +38,7 @@ export function resolveAttackStyles(
 	// use the monsters default attack style.
 	if (
 		attackStyles.length === 0 ||
-		attackStyles.some(s => killableMon.disallowedAttackStyles?.includes(s))
+		attackStyles.some(s => killableMon?.disallowedAttackStyles?.includes(s))
 	) {
 		attackStyles = monsterStyles;
 	}
@@ -65,7 +58,7 @@ export async function addMonsterXP(
 	duration: number
 ) {
 	const [, osjsMon, attackStyles] = resolveAttackStyles(user, monsterID);
-	const hp = osjsMon.data?.hitpoints || miscHpMap[monsterID] || 1;
+	const hp = osjsMon?.data?.hitpoints || miscHpMap[monsterID] || 1;
 	const totalXP = hp * 4 * quantity;
 	const xpPerSkill = totalXP / attackStyles.length;
